Change appId in Identity to Integer

+++ src/main/java/com/enjin/enjincoin/sdk/model/service/identities/Identity.java
@@ -34,7 +34,7 @@ public class Identity {
      * @return the app id.
      */
     @SerializedName("app_id")
-    private BigInteger appId;
+    private Integer appId;
 
     /**
      * The linked ethereum address.
